Module hyperscan::chimera

Chimera is a software regular expression matching engine that is a hybrid of Hyperscan and PCRE.

The design goals of Chimera are to fully support PCRE syntax as well as to take advantage of the high performance nature of Hyperscan.

Examples

let db: Database = "/test/i".parse().unwrap();
let scratch = db.alloc_scratch().unwrap();
let mut matches = vec![];
let mut errors = vec![];

db.scan("some test data", &scratch, |id, from, to, _flags, captured| {
    matches.push((from, to));

    Matching::Continue
}, |error_type, id| {
    errors.push((error_type, id));

    Matching::Skip
}).unwrap();

assert_eq!(matches, vec![(5, 9)]);
assert_eq!(errors, vec![]);

Modules

prelude

The chimera Prelude

Structs

Capture

Structure representing a captured subexpression within a match.

CompileError

Providing details of the compile error condition.

Database

A compiled pattern database that can then be used to scan data.

DatabaseRef

A borrowed reference to a Database.

Flags

Pattern flags

Pattern

The pattern with basic regular expression.

Patterns

Vec of Pattern

Scratch

A large enough region of scratch space to support a given database.

ScratchRef

A borrowed reference to a Scratch.

Enums

Error

A type for errors returned by Chimera functions.

MatchError

The type of error event that occurred.

Matching

Callback return value used to tell the Chimera matcher what to do after processing this match.

Mode

Compile mode flags

Traits

Builder

The regular expression pattern database builder.

ErrorEventHandler

Definition of the Chimera error event callback function type.

MatchEventHandler

Definition of the match event callback function type.

Functions

compile

Compile an expression into a Chimera database.

version

Utility function for identifying this release version.