Enum hyperscan::Error

pub enum Error {
    Invalid,
    NoMem,
    ScanTerminated,
    CompilerError(String),
    DbVersionError,
    DbPlatformError,
    DbModeError,
    BadAlign,
    BadAlloc,
    Failed(i32),
    ParseError(ParseIntError),
    NulError(NulError),
}

Error Codes

Variants
Invalid

A parameter passed to this function was invalid.

NoMem

A memory allocation failed.

ScanTerminated

The engine was terminated by callback.

This return value indicates that the target buffer was partially scanned, but that the callback function requested that scanning cease after a match was located.

CompilerError(String)

The pattern compiler failed with more detail.

DbVersionError

The given database was built for a different version of Hyperscan.

DbPlatformError

The given database was built for a different platform (i.e., CPU type).

DbModeError

The given database was built for a different mode of operation. This error is returned when streaming calls are used with a block or vectored database and vice versa.

BadAlign

A parameter passed to this function was not correctly aligned.

BadAlloc

The memory allocator (either malloc() or the allocator set with hs_set_allocator()) did not correctly return memory suitably aligned for the largest representable data type on this platform.

Failed(i32)

Unknown error code

ParseError(ParseIntError)

An error which can be returned when parsing an integer.

NulError(NulError)

An error returned from CString::new to indicate that a nul byte was found in the vector provided.