Module hyperscan::prelude

The hyperscan Prelude

Re-exports

pub use crate::pattern;

Structs

CompileFlags

Pattern flags

Database

A compiled pattern database that can then be used to scan data.

Pattern

The pattern with basic regular expression.

Patterns

Vec of Pattern

Scratch

A large enough region of scratch space to support a given database.

Stream

A pattern matching state can be maintained across multiple blocks of target data

Enums

Matching

Indicating whether or not matching should continue on the target data.

Traits

Builder

The regular expression pattern database builder.

Mode

Compile mode

Functions

compile

Compile an expression into a Hyperscan database.

Type Definitions

BlockDatabase

Block scan (non-streaming) database.

StreamingDatabase

Streaming database.

VectoredDatabase

Vectored scanning database.