Trait hyperscan::DatabaseBuilder

pub trait DatabaseBuilder {
    type Err;
    fn for_platform<T: Mode>(
        &self, 
        platform: Option<&PlatformRef>
    ) -> Result<Database<T>, Self::Err>;

    fn build<T: Mode>(&self) -> Result<Database<T>, Self::Err> { ... }
}

The regular expression pattern database builder.

Associated Types

type Err

The associated error which can be returned from compiling.

Required methods

fn for_platform<T: Mode>(
&self,
platform: Option<&PlatformRef>
) -> Result<Database<T>, Self::Err>

Build an expression is compiled into a Hyperscan database for a target platform.

Provided methods

fn build<T: Mode>(&self) -> Result<Database<T>, Self::Err>

Build an expression is compiled into a Hyperscan database which can be passed to the runtime functions