Struct hyperscan::RawDatabase

pub struct RawDatabase<T: Type> {
    // some fields omitted
}

A compiled pattern database that can then be used to scan data.

Methods

impl<T: Type> RawDatabase<T>

fn from_raw(db: RawDatabasePtr) -> RawDatabase<T>

Constructs a compiled pattern database from a raw pointer.

fn free(&mut self) -> Result<(), Error>

Free a compiled pattern database.

impl RawDatabase<Streaming>

fn stream_size(&self) -> Result<usize, Error>

Provides the size of the stream state allocated by a single stream opened against the given database.

impl<T: Type> RawDatabase<T>

fn compile(expression: &str, flags: u32, platform: &PlatformInfo) -> Result<RawDatabase<T>, Error>

The basic regular expression compiler.

This is the function call with which an expression is compiled into a Hyperscan database which can be passed to the runtime functions.